+++ lib/domain/device/service_repository_type.dart
@@ -1,4 +1,4 @@
-import 'package:upnp_explorer/infrastructure/upnp/service_description.dart';
+import '../../infrastructure/upnp/service_description.dart';
 
 abstract class ServiceRepositoryType {
   ServiceDescription? get(String id);