+++ lib/infrastructure/settings/options_repository.dart
@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:injectable/injectable.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
-import '../../domain/value_converter.dart';
 import '../../application/settings/options.dart';
+import '../../domain/value_converter.dart';
 
 const _kThemeKey = 'ThemeMode';
 const _kVisualDensityHorizontal = 'visualDensity_horizontal';