+++ lib/presentation/device/widgets/device-list-item.dart
@@ -1,62 +1,48 @@
 import 'package:flutter/material.dart';
 
+import '../../../application/application.dart';
 import '../../../application/l10n/generated/l10n.dart';
+import '../../../application/routing/routes.dart';
 import '../../../domain/device/device.dart';
+import '../pages/device_page.dart';
 import 'device-image.dart';
 
-class DeviceListItem extends StatefulWidget {
+class DeviceListItem extends StatelessWidget {
   final UPnPDevice device;
-  final void Function(UPnPDevice) onTap;
 
   const DeviceListItem({
     Key? key,
     required this.device,
-    required this.onTap,
   }) : super(key: key);
 
   @override
-  _DeviceListItemState createState() => _DeviceListItemState();
-}
-
-class _DeviceListItemState extends State<DeviceListItem>
-    with SingleTickerProviderStateMixin {
-  late AnimationController _controller;
-  late Animation<double> _animation;
-
-  @override
-  void initState() {
-    _controller = AnimationController(
-      vsync: this,
-      duration: Duration(milliseconds: 250),
-    );
-
-    _animation = Tween(begin: 0.0, end: 1.0).animate(_controller);
-
-    super.initState();
-  }
-
-  @override
   Widget build(BuildContext context) {
-    _controller.forward();
-
-    return FadeTransition(
-      opacity: _animation,
-      child: Semantics(
-        label: S
-            .of(context)
-            .discoveredDevice(widget.device.description.device.friendlyName),
-        child: ListTile(
-          leading: DeviceImage(
-            icons: widget.device.description.device.iconList.icons,
-            deviceIp: widget.device.ipAddress,
-          ),
-          title: Text(
-            widget.device.description.device.friendlyName,
-          ),
-          subtitle: Text(widget.device.ipAddress.toString()),
-          trailing: Icon(Icons.chevron_right),
-          onTap: () => widget.onTap(widget.device),
+    return Semantics(
+      label: S
+          .of(context)
+          .discoveredDevice(device.description.device.friendlyName),
+      child: ListTile(
+        leading: DeviceImage(
+          icons: device.description.device.iconList.icons,
+          deviceIp: device.ipAddress,
         ),
+        title: Text(
+          device.description.device.friendlyName,
+        ),
+        subtitle: Text(device.ipAddress.toString()),
+        trailing: Icon(Icons.chevron_right),
+        onTap: () {
+          Application.router!.navigateTo(
+            context,
+            Routes.deviceDocument,
+            routeSettings: RouteSettings(
+              arguments: DevicePageArguments(
+                device.description.device,
+                device.discoveryResponse,
+              ),
+            ),
+          );
+        },
       ),
     );
   }