+++ CHANGELOG.md
@@ -4,7 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## Unreleased
+### Changed
+- Animations on page transitions.
+- Animations on device discovered.
+- Moved scanning indicator to bottom of screen to prevent the devices list from moving when scanning is complete.
+
+## [1.0.10] - 2022-09-11
 ### Added
 - Changelog dialog.
 - App review button.